+++ src/main/java/salute/oneshot/domain/ingredient/repository/IngredientRepository.java
@@ -4,4 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import salute.oneshot.domain.ingredient.entity.Ingredient;
 
 public interface IngredientRepository extends JpaRepository<Ingredient, Long> {
+
+
 }
