+++ src/main/java/salute/oneshot/domain/ingredient/service/IngredientService.java
@@ -27,6 +27,7 @@ import salute.oneshot.domain.ingredient.entity.IngredientDocument;
 
 import salute.oneshot.domain.ingredient.repository.IngredientElasticSearchRepository;
 import salute.oneshot.domain.ingredient.repository.IngredientRepository;
+import salute.oneshot.domain.ingredient.repository.IngredientSearchFinder;
 import salute.oneshot.global.exception.NotFoundException;
 import salute.oneshot.global.util.S3Util;
 
@@ -37,10 +38,8 @@ public class IngredientService {
 
     private final IngredientRepository ingredientRepository;
     private final IngredientElasticSearchRepository elasticRepository;
-    private final ElasticsearchClient client;
+    private final IngredientSearchFinder searchFinder;
     private final S3Util s3Util;
-
-    private final String INGREDIENT_INDEX = "ingredients";
 
 
     @Transactional
@@ -109,31 +108,9 @@ public class IngredientService {
     @Transactional(readOnly = true)
     public Page<IngrResponseDto> searchByCondition(SearchIngrSDto sDto) throws IOException {
 
-        int size = sDto.getPageable().getPageSize();
-        int page = sDto.getPageable().getPageNumber();
-        int from = size * page;
+        SearchResponse<IngredientDocument> searchResponse = searchFinder.ingrSearchByCondition(sDto);
 
-
-        BoolQuery.Builder builder = QueryBuilders.bool();
-
-        if (!sDto.getKeyword().isBlank()) {
-            addShouldIfNotNull(builder, sDto.getKeyword(), "name", 3.0f);
-            addShouldIfNotNull(builder, sDto.getKeyword(), "description", 2.0f);
-        }
-
-        if (sDto.getCategory() != null) {
-            addShouldIfNotNull(builder, sDto.getCategory(), "category", 1.0f);
-        }
-
-        SearchRequest searchRequest = new SearchRequest.Builder()
-                .index(INGREDIENT_INDEX)
-                .from(from)
-                .size(size)
-                .query(q -> q.bool(builder.build())).build();
-
-        SearchResponse<IngredientDocument> response = client.search(searchRequest, IngredientDocument.class);
-
-        Map<Long, Integer> responseInr = response.hits().hits().stream()
+        Map<Long, Integer> responseInr = searchResponse.hits().hits().stream()
                 .filter(hit -> hit.source() != null)
                 .collect(Collectors.toMap(
 
@@ -148,16 +125,11 @@ public class IngredientService {
                         Comparator.reverseOrder()))
                 .toList();
 
-        long total = response.hits().total() == null ? 0 : response.hits().total().value();
+        long total = searchResponse.hits().total() == null ? 0 : searchResponse.hits().total().value();
 
         return new PageImpl<>(ingredientList, sDto.getPageable(), total);
     }
 
-
-    private void addShouldIfNotNull(BoolQuery.Builder builder, String condition, String fieldName, float boost) {
-        builder.should(Query.of(q -> q.match(m -> m.field(fieldName)
-                .query(condition.toLowerCase()).boost(boost))));
-    }
 
     private Ingredient findById(Long id) {
         return ingredientRepository.findById(id)
